Loota shit

+++ server/order-service/src/main/java/com/alexandre/orderservice/controller/OrderController.java
@@ -1,13 +1,16 @@
 package com.alexandre.orderservice.controller;
 
 import com.alexandre.orderservice.dto.order.OrderDTO;
+import com.alexandre.orderservice.exception.NotFoundException;
+import com.alexandre.orderservice.record.UserPrincipal;
 import com.alexandre.orderservice.service.order.OrderService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 import java.util.UUID;
 
 @RequiredArgsConstructor
@@ -17,23 +20,36 @@ public class OrderController {
     private final OrderService orderService;
 
     @PostMapping("/")
-    public ResponseEntity<OrderDTO> createOrder(@RequestBody @Valid OrderDTO orderDTO) {
+    public ResponseEntity<OrderDTO> createOrder(@AuthenticationPrincipal UserPrincipal userPrincipal, @RequestBody @Valid OrderDTO orderDTO) {
+        orderDTO.setProfileId(userPrincipal.profile().getId());
         return ResponseEntity.ok(orderService.create(orderDTO));
     }
 
     @PutMapping("/")
-    public ResponseEntity<OrderDTO> updateOrder(@RequestBody @Valid OrderDTO orderDTO) {
-        return ResponseEntity.ok(orderService.update(orderDTO));
+    public ResponseEntity<OrderDTO> updateOrder(@AuthenticationPrincipal UserPrincipal userPrincipal, @RequestBody @Valid OrderDTO orderDTO) {
+        if (orderService.profileOwnsOrder(orderDTO.getId(), userPrincipal.profile().getId())) {
+            return ResponseEntity.ok(orderService.update(orderDTO));
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<OrderDTO> getOrder(@PathVariable UUID id) {
-        return ResponseEntity.ok(orderService.findById(id));
+    public ResponseEntity<OrderDTO> getOrder(@AuthenticationPrincipal UserPrincipal userPrincipal, @PathVariable UUID id) {
+        if (orderService.profileOwnsOrder(id, userPrincipal.profile().getId())) {
+            return ResponseEntity.ok(orderService.findById(id));
+        } else {
+            throw new NotFoundException("Order not found");
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable UUID id) {
-        orderService.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Void> deleteOrder(@AuthenticationPrincipal UserPrincipal userPrincipal, @PathVariable UUID id) {
+        if (orderService.profileOwnsOrder(id, userPrincipal.profile().getId())) {
+            orderService.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            throw new NotFoundException("Order not found");
+        }
     }
 }
